Tests/LibWeb/Text/expected/HTML/storage-does-not-have-legacy-override-builtins-flag.txt

key should still be native: 'function key() { [native code] }'
key's stringified function was added to storage: 'function () { println(`FAIL: Overriden ${functionName} was called`); }'
getItem should still be native: 'function getItem() { [native code] }'
getItem's stringified function was added to storage: 'function () { println(`FAIL: Overriden ${functionName} was called`); }'
setItem should still be native: 'function setItem() { [native code] }'
setItem's stringified function was added to storage: 'function () { println(`FAIL: Overriden ${functionName} was called`); }'
removeItem should still be native: 'function removeItem() { [native code] }'
removeItem's stringified function was added to storage: 'function () { println(`FAIL: Overriden ${functionName} was called`); }'
clear should still be native: 'function clear() { [native code] }'
clear's stringified function was added to storage: 'function () { println(`FAIL: Overriden ${functionName} was called`); }'
